add dry run and refactoring
