+++ src/RepairManagerAgent/commands.py
@@ -1,0 +1,5 @@
+import os
+
+def reboot_node():
+    os.system('sync')
+    os.system('reboot -f')
